Separate modal class for reuse in other classes
